keyboard_layout: split drivers into separate files
